refactor: split the product model out of the Shopify fetch layer (#68)
Building with the flag unset put the whole development catalogue into
.next/static — the CLIENT bundle. Fourteen fake products downloaded by every
visitor. Never rendered, but shipped, and directly against the requirement that
development data stay out of production.

Cause: four "use client" components imported from lib/shopify — the card, the
store grid, the featured grid and the detail page — so the fetch layer was
already in the client graph. With the flag set to "false" the guarded import()
is statically dead and gets dropped, which is why the earlier check looked
clean. With it UNSET the comparison can't be folded, the branch stays
reachable, and the catalogue travels with it. Unset is the documented safe
default and what a fresh clone has, so that was the common case.

Splits the module along the seam the brief already asked for — keep data
fetching separate from presentation:

  lib/product.ts  the shape plus pure helpers (isSoldOut, toCartItem,
                  defaultVariant, getRelatedProducts, displayTags,
                  getShowcaseProducts). Safe anywhere.
  lib/shopify.ts  Storefront queries, toProduct, the fetchers and the catalogue
                  switch. Server only.

Client components now import lib/product and can no longer reach the queries or
the catalogue behind them. Verified: unset builds ship 0 bytes of it to the
client. It remains in the server graph when unset, unreachable behind a false
condition; setting the flag explicitly to false removes it there too, which is
what .env.example recommends for production.

Also drops the shopify -> placeholder-products -> shopify import cycle, since
FEATURED_TAG now lives in the leaf module.
