+++ src/lib/shopify.ts
@@ -1,7 +1,14 @@
-// Catalog data layer. Every storefront surface (home featured grid, store grid,
-// product detail) reads products from here, so Shopify is the single source of
-// truth. When the store isn't configured we fall back to non-purchasable
-// placeholders so the site still renders.
+// Catalog data layer — Storefront API fetching. Every storefront surface (home
+// featured grid, store grid, product detail) reads products from here, so Shopify
+// is the single source of truth. When the store isn't configured we fall back to
+// non-purchasable placeholders so the site still renders.
+//
+// SERVER ONLY. The product shape and the pure helpers live in `./product`, which
+// is what client components import. Keeping them apart matters: this module
+// reaches the Storefront queries and the development catalogue, and a client
+// component importing it drags all of that into the browser bundle (#68).
+
+import type { ShopifyProduct, ShopifyVariant } from "./product";
 
 const PRODUCT_FRAGMENT = `
   fragment CatalogProduct on Product {
@@ -81,78 +88,6 @@ interface ShopifyProductNode {
   };
   rating: { value: string } | null;
   ratingCount: { value: string } | null;
-}
-
-export interface ShopifyVariant {
-  /** Numeric Shopify variant id — the `#<variantId>` half of the cart key. */
-  id: string;
-  label: string;
-  price: number;
-  /**
-   * This variant's own compare-at price, unlike `ShopifyProduct.originalPrice`
-   * which is the product-level range and only lines up with the cheapest variant.
-   */
-  originalPrice: number | null;
-  availableForSale: boolean;
-}
-
-export interface ShopifyProduct {
-  /** Namespaced cart key, e.g. "product:1" — never a bare number (see CartContext). */
-  id: string;
-  name: string;
-  category: string;
-  price: number;
-  originalPrice: number | null;
-  image: string;
-  tags: string[];
-  /** Placeholders aren't purchasable: no cart, no wishlist, no detail page. */
-  placeholder?: boolean;
-  /** Shopify handle; enables the card's link to `/store/<slug>`. */
-  slug?: string;
-  description?: string;
-  gallery?: string[];
-  variants?: ShopifyVariant[];
-  /** Only present when the store publishes the `reviews` metafields. */
-  rating?: number;
-  reviews?: number;
-  /** Explicit badge label; falls back to a "Sale" badge when `originalPrice` is set. */
-  badge?: string;
-  badgeColor?: string;
-}
-
-/**
- * Marks a product for the store's featured showcase. A plain Shopify tag, so
- * merchandisers control the showcase from admin without a deploy, and the
- * development catalogue marks its picks the same way — one mechanism, both
- * sources. Filtered out of the tags a card displays (see `displayTags`).
- */
-export const FEATURED_TAG = "featured";
-
-export function isFeatured(product: ShopifyProduct): boolean {
-  return product.tags.some((t) => t.toLowerCase() === FEATURED_TAG);
-}
-
-/** Tags a card should render: everything except the merchandising marker. */
-export function displayTags(product: ShopifyProduct): string[] {
-  return product.tags.filter((t) => t.toLowerCase() !== FEATURED_TAG).slice(0, 3);
-}
-
-/**
- * The store's featured showcase, in a stable order — never random, so the same
- * catalogue always yields the same shelf.
- *
- * Tagged products come first. If there aren't enough, it tops up from the rest in
- * catalogue order, which for Shopify is `sortKey: BEST_SELLING` — a real
- * merchandising signal rather than an arbitrary API ordering.
- */
-export function getShowcaseProducts(
-  all: ShopifyProduct[],
-  count = 4
-): ShopifyProduct[] {
-  const tagged = all.filter(isFeatured);
-  if (tagged.length >= count) return tagged.slice(0, count);
-  const rest = all.filter((p) => !isFeatured(p));
-  return [...tagged, ...rest].slice(0, count);
 }
 
 /**
@@ -346,49 +281,4 @@ export async function getProductBySlug(
   );
 
   return data?.product ? toProduct(data.product) : null;
-}
-
-/** Same-category products first, then the rest, capped at 4. */
-export function getRelatedProducts(
-  product: ShopifyProduct,
-  all: ShopifyProduct[]
-): ShopifyProduct[] {
-  const others = all.filter((p) => p.id !== product.id && !p.placeholder);
-  return others
-    .filter((p) => p.category === product.category)
-    .concat(others.filter((p) => p.category !== product.category))
-    .slice(0, 4);
-}
-
-/** The variant a card's quick-add should use: first in stock, else the first. */
-export function defaultVariant(
-  product: ShopifyProduct
-): ShopifyVariant | undefined {
-  return (
-    product.variants?.find((v) => v.availableForSale) ?? product.variants?.[0]
-  );
-}
-
-export function isSoldOut(product: ShopifyProduct): boolean {
-  const variants = product.variants;
-  // A product with no variant data (e.g. a placeholder) isn't "sold out".
-  return !!variants?.length && variants.every((v) => !v.availableForSale);
-}
-
-/**
- * Builds a cart line for a product/variant pair. Variant-specific lines append
- * `#<variantId>` to the product key so two sizes of the same product are
- * distinct rows in the bag (see CartContext).
- */
-export function toCartItem(product: ShopifyProduct, variant?: ShopifyVariant) {
-  const v = variant ?? defaultVariant(product);
-  const hasChoice = (product.variants?.length ?? 0) > 1;
-
-  return {
-    id: v ? `${product.id}#${v.id}` : product.id,
-    name: hasChoice && v ? `${product.name} — ${v.label}` : product.name,
-    category: product.category,
-    price: v?.price ?? product.price,
-    image: product.image,
-  };
 }
